# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Family Manager BI Group API",
        "version": "1.1.0"
    },
    "paths": {
...
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "description": "Get list tasks",
                "operationId": "App\\Http\\Controllers\\TaskController::index",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date_created_from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_created_to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_updated_from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_updated_to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_hash",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "List limit",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 10
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Offset",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 0
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return sections list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "USER_HAS_NO_PERMISSION"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "operationId": "App\\Http\\Controllers\\TaskController::store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return task data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskRequest"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "USER_HAS_NO_PERMISSION"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "description": "Get single task",
                "operationId": "App\\Http\\Controllers\\TaskController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return task data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskRequest"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "USER_HAS_NO_PERMISSION"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "description": "Update task",
                "operationId": "App\\Http\\Controllers\\TaskController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return Task model"
                    },
                    "403": {
                        "description": "USER_HAS_NO_PERMISSION"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tasks/export": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "description": "Export tasks",
                "operationId": "App\\Http\\Controllers\\TaskController::export",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date_created_from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_created_to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_updated_from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_updated_to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_hash",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return tasks excel"
                    },
                    "403": {
                        "description": "USER_HAS_NO_PERMISSION"
                    }
                }
            }
        },
        "/api/tasks/creators": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "description": "Tasks creators",
                "operationId": "App\\Http\\Controllers\\TaskController::creators",
                "responses": {
                    "200": {
                        "description": "Return tasks creators"
                    },
                    "403": {
                        "description": "USER_HAS_NO_PERMISSION"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tasks/executors": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "description": "Tasks executors",
                "operationId": "App\\Http\\Controllers\\TaskController::executors",
                "responses": {
                    "200": {
                        "description": "Return tasks executors"
                    },
                    "403": {
                        "description": "USER_HAS_NO_PERMISSION"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TaskRequest": {
                "title": "Task request",
                "description": "Task request body data",
                "required": [
                    "title",
                    "content"
                ],
                "properties": {
                    "title": {
                        "schema": "TaskRequest",
                        "type": "string",
                        "example": "Заголовок"
                    },
                    "content": {
                        "schema": "TaskRequest",
                        "type": "string",
                        "example": "Содержимое заявки/задачи. Текст большого объёма."
                    }
                },
                "type": "object"
            },
...
            "TaskUpdateRequest": {
                "title": "Task update request",
                "description": "Task request body data",
                "required": [
                    "title",
                    "content"
                ],
                "properties": {
                    "title": {
                        "schema": "TaskUpdateRequest",
                        "type": "string",
                        "example": "Заголовок"
                    },
                    "content": {
                        "schema": "TaskUpdateRequest",
                        "type": "string",
                        "example": "Содержимое заявки/задачи. Текст большого объёма."
                    }
                },
                "type": "object"
            },
...
            "TaskResource": {
                "title": "TaskResource",
                "description": "Task resource",
                "properties": {
                    "id": {
                        "schema": "TaskResource",
                        "type": "integer"
                    },
                    "status_id": {
                        "schema": "TaskResource",
                        "type": "integer"
                    },
                    "title": {
                        "schema": "TaskResource",
                        "type": "string"
                    },
                    "content": {
                        "schema": "TaskResource",
                        "type": "string"
                    },
                    "author": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "created_at": {
                        "schema": "TaskResource",
                        "type": "string"
                    },
                    "updated_at": {
                        "schema": "TaskResource",
                        "type": "string"
                    },
                    "deleted_at": {
                        "schema": "TaskResource",
                        "type": "string"
                    },
                    "executor": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "comment": {
                        "schema": "TaskResource",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "title": "UserResource",
                "description": "User resource",
                "properties": {
                    "user_hash": {
                        "schema": "UserResource",
                        "type": "string"
                    },
                    "firstName": {
                        "schema": "UserResource",
                        "type": "string"
                    },
                    "lastName": {
                        "schema": "UserResource",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}